refactor(nav): move user menu into a composable
